% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sbp_visits.R
\name{avg_sbp_visits}
\alias{avg_sbp_visits}
\title{Average Systolic Blood Pressures and Visit Information}
\usage{
avg_sbp_visits(df, birth_date_var, pat_mrn, pcp_var, enc_date_var, sbp_var,
  dbp_var, age_var)
}
\arguments{
\item{df}{EHR dataset which has had dates properly formatted by format_ehr_date function}

\item{birth_date_var}{The variable name which includes information pertaining to birthdate}

\item{pat_mrn}{The variable name which includes the Patient MRN information}

\item{pcp_var}{The variable name which includes the names of the PCP's}

\item{enc_date_var}{The variable name which includes information pertaining to visit/encounter dates}

\item{sbp_var}{The variable name which includes information pertaining to systolic blood pressure}

\item{dbp_var}{The variable name which includes information pertaining to diastolic blood pressure}

\item{age_var}{The variable name which includes information pertaining to age}
}
\value{
A data frame with several newly derived variables related to sbp and number of visits
}
\description{
This function allows you to find average SBP, DBP, and number of visits per month for patients
    in addition to calculating total number of visits.
}
\examples{
\dontrun{
average_sbp(df, birth_date_var = Birth.Date, pat_mrn = Patient.MRN, pcp_var = Patient.PCP,
enc_date_var = Encounter.Date, sbp_var = Systolic.BP, dbp_var = Diastolic.BP,
age_var = age_at_last_encounter)
}
}
