% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate Age of Patients}
\usage{
calculate_age(df, birth_date_var, end_date_var, new_var_name)
}
\arguments{
\item{df}{Raw EHR dataset which includes birthdate covariate}

\item{birth_date_var}{Birthdate covariate}

\item{end_date_var}{End date for age calculation}

\item{new_var_name}{What the newly created age variable should be called}
}
\value{
A data frame with a derived age variable
}
\description{
This function calculates ages from patient birthdays
}
\examples{
\dontrun{
calculate_age(df, Birth.Date, Encounter.Date, Derived.Age)
}
}
