% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_patients.R
\name{find_patients}
\alias{find_patients}
\title{Find Patients who see PCP's of interest within past year}
\usage{
find_patients(df, cutoff_date, vec_of_pcps, pcp_var, pat_mrn, enc_date_var)
}
\arguments{
\item{df}{EHR dataset which has had dates properly formatted by format_ehr_date function}

\item{cutoff_date}{End date of interest in month-day-4digityear format (ie. 11012018 is Nov 1, 2018)}

\item{vec_of_pcps}{A vector listing names of PCP's in format found in EHR}

\item{pcp_var}{The variable name which includes the names of the PCP's}

\item{pat_mrn}{The variable name which includes the Patient MRN information}

\item{enc_date_var}{The variable name which includes information pertaining to visit/encounter dates}
}
\value{
A subsetted data frame which includes patients with PCP's of interest and within date range
}
\description{
This function allows you to find the patients who see PCP's of interest and who have had visits
    within one year of the cutoff date.
}
\examples{
\dontrun{
find_patients(df, cutoff_date, vec_of_pcps = pcp_vec, pcp_var = Patient.PCP,
pat_mrn = Patient.MRN, enc_date_var = Encounter.Date)
}
}
