% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ehr_date.R
\name{format_ehr_date}
\alias{format_ehr_date}
\title{Format EHR Date}
\usage{
format_ehr_date(df, covariate)
}
\arguments{
\item{df}{Raw EHR dataset which includes covariate}

\item{covariate}{Date covariate which needs to be reformatted}
}
\value{
A data frame with a reformatted date
}
\description{
This function allows you to format dates pulled from WFBMC EHR into standard format
}
\examples{
\dontrun{
format_ehr_date(df, Birth.Date)
}
}
