% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_seen_nephrologist.R
\name{has_seen_nephrologist}
\alias{has_seen_nephrologist}
\title{Find patients who have seen a nephrologist within specified time period and create flag}
\usage{
has_seen_nephrologist(df, cutoff_date, num_years, enc_date_var, pat_mrn)
}
\arguments{
\item{df}{Raw EHR dataset which includes blood pressure data}

\item{cutoff_date}{End date of interest in month-day-4digityear format (ie. 11012018 is Nov 1, 2018)}

\item{num_years}{Number of years to search back from cutoff_date}

\item{pat_mrn}{The variable name which includes the Patient MRN information}
}
\value{
A data table listing Patient MRN and an indicator variable for nephrology visit
}
\description{
This function allows you to create a dataframe which includes the Patient MRN
    and an indicator variable (returns 'Yes' for 'has_seen_neph') for all patients who have seen
    a nephrologist within a specified time period.  This dataframe can then be merged with a summarized
    blood pressure data table (see wfbmcr::summarize_bp_dt' for data table creation and
    'wfbmcr::merge_w_indicator' for merging)
}
\examples{
\dontrun{
has_seen_nephrologist(df, enc_date_var = CONTACT_DATE, cutoff_date, num_years = 1,
pat_mrn = PAT_MRN_ID)
}
}
