% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nested_list.R
\name{make_nested_list}
\alias{make_nested_list}
\title{Function to make nested lists (store modeling output)}
\usage{
make_nested_list(dim_lvl_1, dim_lvl_2)
}
\arguments{
\item{dim_lvl_1}{Dimensions of outer level (ie. Number of PCP's)}

\item{dim_lvl_2}{Dimensions of inner level (ie. Number of covariates)}
}
\value{
A nested list with two levels
}
\description{
Function to make nested lists (store modeling output)
}
\examples{
\dontrun{
make_nested_list(5, 7)
}
}
