% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbp_lsmeans.R
\name{sbp_lsmeans}
\alias{sbp_lsmeans}
\title{Calculate LS Means of SBP for Spatial Correlation Model}
\usage{
sbp_lsmeans(df)
}
\arguments{
\item{df}{EHR dataset which has had variable names properly formatted to match modeling}
}
\value{
A data frame with lsmeans for avg_sbp by month and year
}
\description{
This function calculates a spatially correlated covariance matrix and then finds the LS means for
    SBP from an EHR dataset which has the following names: avg_sbp, sbp_calc_date, and Patient.MRN
}
\examples{
\dontrun{
sbp_lsmeans(df)
}
}
\seealso{
\code{\link[nlme]{gls}}, \code{\link[emmeans]{emmeans}}
}
