% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bp_dt.R
\name{summarize_bp_dt}
\alias{summarize_bp_dt}
\title{Summarize Blood Pressure Data Table}
\usage{
summarize_bp_dt(df, cutoff_date, num_years, enc_date_var, pat_mrn, sbp_var,
  enc_csn_var, pcp_var, vec_of_pcps)
}
\arguments{
\item{df}{Raw EHR dataset which includes blood pressure data}

\item{cutoff_date}{End date of interest in month-day-4digityear format (ie. 11012018 is Nov 1, 2018)}

\item{num_years}{Number of years to search back from cutoff_date}

\item{enc_date_var}{The variable name which includes information pertaining to visit/encounter dates}

\item{pat_mrn}{The variable name which includes the Patient MRN information}

\item{sbp_var}{The variable name which includes information pertaining to systolic blood pressure}

\item{enc_csn_var}{The variable name which includes information pertaining to encounter CSN}

\item{pcp_var}{The variable name which includes the names of the PCP's}

\item{vec_of_pcps}{A vector listing names of PCP's in format found in EHR}
}
\value{
A data table summarizing patient blood pressure information
}
\description{
This function allows you summarize a dataframe consisting of blood pressure data and
    put it into a data table
}
\examples{
\dontrun{
summarize_bp_dt(df, cutoff_date, num_years = 1, vec_of_pcps, enc_date_var = Encounter.Date,
 sbp_var = Systolic.BP, enc_csn_var = Encounter.CSN, pat_mrn = Patient.MRN, pcp_var = Patient.PCP)
}
}
